requirements.txt
Create two virtual environments, and install the respective dependencies in each
.venv_vision
    openai-whisper  
    numpy
    opencv
    mediapipe (8)
    #For macOS, brew install portaudio first before pip install pyaudio
    pyaudio

.venv_agent
    google-adk (9)
